+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - #130 Closes Find prompt when opening a file
 - #129 Fixes FileSystem Move and Copy to handle empty folders
 - Fixes Find input error not going away after using Go To Next Find Match
+- Fixes drive not being selected when opening parent
 - Various optimizations for file write operations
 
 ### 0.6.2  (2022-12-23)